Implemented rating calculation and storage on the back end
Whenever a new high score gets stored, the rating for that high score gets calculated and stored. If a user hasn't played a level to completion yet, there is no rating stored and a query asking for one will return null (as opposed to a rating of 0 if they have played but performed poorly).