Updated the Star fields to have an ellipses placeholder while fetching the database values
So the page can load faster, the fetching of the rating totals is done asynchronously. As a result, the values would just pop into existence after half a second or so. To lessen the oddity of this and communicate to the user that the value is loading, a placeholder ("...") is put in immediately and replaced once the value reaches the local browser.