Fixed Custom Session Permissions Error and Added Custom Session Automated Test
Previously, custom session had 2 permission errors.
First, it did not check to make sure a user was logged in, meaning someone without an account could access a custom session as long as they had the URL.
Second, users who were logged in properly could not have their results displayed because the client would request that the database update the highscore and associated rating for a non-MainGame level.