Updated Existing Tests to implicitly_wait after each attempted driver action
Previously the tests used python's time.sleep() method to have the driver wait a set amount of time between each command, but this was not a good way of accounting for latency as if the set time was too large time would be wasted doing nothing but if it was too small then a lag spike could error out the test. After some research, I discovered the implicitly_wait() method selenium has which allows for the driver to repeatedly poll for an action completion or element on the page. Some spots still use time.sleep() if instantly inputting commands/keys is too fast for the web scripts to handle, and an extra second is needed to buffer those spots out.