Added more User Tests, as well as a batch file which executes all tests in order
To ensure the same starting point for the tests each run, the tests are executed in an order such that a new user is tested each time for the user tests. This way the process of "unlocking" new levels can be observed, as well as the unique steps a user only sees during their first login.