Updated the deleteUser google cloud function to cascade deletion to a user's highscore and rating total subcollections
The previous deletion implementation didn't account for the newly added subcollections, and as a result even though the parent user document was being deleted, the children subcollections remained. Even though these are comparable to objects with no references pointing to them and wouldn't cause issues with functionality on the front-end, for garbage collection this needed to be updated to avoid the database becoming filled with deleted user data remnants.